+++ sarif/__init__.py
@@ -2,14 +2,4 @@
 Top-level version information for sarif-tools.
 """
 
-import importlib.metadata
-
-
-def _read_package_version():
-    try:
-        return importlib.metadata.version("sarif-tools")
-    except importlib.metadata.PackageNotFoundError:
-        return "local"
-
-
-__version__ = _read_package_version()
+__version__ = "2.1.0"