+++ sarif/operations/summary_op.py
@@ -5,12 +5,11 @@ Code for `sarif summary` command.
 import os
 from typing import List
 
-from sarif import sarif_file
 from sarif.sarif_file import SarifFileSet
 
 
 def generate_summary(
-    input_files: SarifFileSet, output: str, output_multiple_files: bool
+    input_files: SarifFileSet, output: str, output_multiple_files: bool, code_only: bool
 ):
     """
     Generate a summary of the issues from the SARIF files.
@@ -25,7 +24,7 @@ def generate_summary(
                 input_file.get_file_name_without_extension() + "_summary.txt"
             )
             output_file = os.path.join(output, output_file_name)
-            summary_lines = _generate_summary(input_file)
+            summary_lines = _generate_summary(input_file, code_only)
             print(
                 "Writing summary of",
                 input_file.get_file_name(),
@@ -37,7 +36,7 @@ def generate_summary(
         output_file_name = "static_analysis_summary.txt"
         output_file = os.path.join(output, output_file_name)
 
-    summary_lines = _generate_summary(input_files)
+    summary_lines = _generate_summary(input_files, code_only)
     if output:
         print(
             "Writing summary of",
@@ -52,7 +51,7 @@ def generate_summary(
             print(lstr)
 
 
-def _generate_summary(input_files: SarifFileSet) -> List[str]:
+def _generate_summary(input_files: SarifFileSet, code_only: bool) -> List[str]:
     """
     For each severity level (in priority order): create a list of the errors of
     that severity, print out how many there are and then do some further analysis
@@ -61,10 +60,14 @@ def _generate_summary(input_files: SarifFileSet) -> List[str]:
     ret = []
     result_count_by_severity = input_files.get_result_count_by_severity()
     for severity in input_files.get_severities():
-        issue_code_histogram = input_files.get_issue_code_histogram(severity)
+        issue_cd_histogram = (
+            input_files.get_issue_code_histogram(severity)
+            if code_only
+            else input_files.get_issue_code_and_description_histogram(severity)
+        )
         result_count = result_count_by_severity.get(severity, 0)
         ret.append(f"\n{severity}: {result_count}")
-        ret += [f" - {code}: {count}" for (code, count) in issue_code_histogram]
+        ret += [f" - {code}: {count}" for (code, count) in issue_cd_histogram]
     filter_stats = input_files.get_filter_stats()
     if filter_stats:
         ret.append(f"\nResults were filtered by {filter_stats}")